updated collision_free_event method in probabilistic_action.py, removed it from parameterizer.py of pycram
