+++ pycram/src/pycram/designators/specialized_designators/probabilistic/probabilistic_action.py
@@ -2,6 +2,7 @@ import enum
 from dataclasses import dataclass
 from typing import Type
 
+import numpy as np
 from probabilistic_model.distributions import SymbolicDistribution
 from probabilistic_model.probabilistic_circuit.rx.helper import fully_factorized, leaf
 from probabilistic_model.probabilistic_circuit.rx.probabilistic_circuit import (
@@ -10,12 +11,19 @@ from probabilistic_model.probabilistic_circuit.rx.probabilistic_circuit import (
     ProductUnit,
 )
 from probabilistic_model.utils import MissingDict
-from random_events.product_algebra import SimpleEvent
+from random_events.interval import singleton
+from random_events.product_algebra import SimpleEvent, Event
 from random_events.set import Set
 from random_events.variable import Symbolic, Continuous
+
+
+from sortedcontainers import SortedSet
+
+from semantic_digital_twin.datastructures.variables import SpatialVariables
 from semantic_digital_twin.spatial_types import HomogeneousTransformationMatrix
 from semantic_digital_twin.world import World
 from semantic_digital_twin.world_description.geometry import BoundingBox
+from semantic_digital_twin.world_description.graph_of_convex_sets import GraphOfConvexSets
 from semantic_digital_twin.world_description.shape_collection import (
     BoundingBoxCollection,
 )
@@ -28,7 +36,6 @@ from ....datastructures.enums import Arms, Grasp, VerticalAlignment, ApproachDir
 from ....datastructures.grasp import GraspDescription
 from ....datastructures.partial_designator import PartialDesignator
 from ....datastructures.pose import PoseStamped
-from ....parameterizer import collision_free_event
 from ....utils import classproperty
 
 
@@ -113,8 +120,54 @@ class MoveAndPickUpParameterizer(ProbabilisticAction):
         bb_collection = BoundingBoxCollection(
             [search_space], reference_frame=self.world.root
         )
-        navigate_conditions = collision_free_event(obj._world, bb_collection)
+        navigate_conditions = self.collision_free_event(obj._world, bb_collection)
         return navigate_conditions
+
+    def collision_free_event(
+            world: World, search_space: Optional[BoundingBoxCollection] = None
+    ) -> Event:
+        """
+        Create an event that describes the free space of the world.
+        :param world: The world to create the event from.
+        :param search_space: The search space to limit the collision free event to.
+        :return: An event that describes the free space.
+        """
+
+        # xy = SortedSet([BoundingBox.x_variable, BoundingBox.y_variable])
+        xy = SpatialVariables.xy
+
+        # create search space for calculations
+        if search_space is None:
+            search_space = BoundingBoxCollection(
+                [
+                    BoundingBox(
+                        -np.inf,
+                        -np.inf,
+                        -np.inf,
+                        np.inf,
+                        np.inf,
+                        np.inf,
+                        origin=HomogeneousTransformationMatrix(reference_frame=world.root),
+                    )
+                ],
+            )
+
+        # remove the z axis
+        search_event = search_space.event
+
+        # get obstacles
+        obstacles = GraphOfConvexSets.obstacles_from_world(world, search_space)
+
+        free_space = search_event - obstacles
+        free_space = free_space.marginal(xy)
+
+        # create floor level
+        z_event = SimpleEvent({SpatialVariables.z.value: singleton(0.0)}).as_composite_set()
+        z_event.fill_missing_variables(xy)
+        free_space.fill_missing_variables(SortedSet([SpatialVariables.z.value]))
+        free_space &= z_event
+
+        return free_space
 
     def accessing_distribution_for_object(
         self, obj: Body, object_variable: Symbolic
